Dealing with slash in subscription name for terraform and MS quirkiness
Azure allow / in the sub name however the API endpoint return wrongly even we
encode the / into %2F. Get
https://management.azure.com/subscriptions/CALM-WFM-Development-Dev%2FTest/providers?api-version=2016-02-01
into your browser and see, azure will return sub not found and the sub name is
the first part of the slash only.

However ID seems to work.

This changes put some more debugging and try to parse and use the ID if found a slash.